# reports/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Demo Target API",
    "version": "0.1.0"
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "用户登录",
        "description": "登录接口精确返回契约（设计断言时必须遵守）：\n1) 仅当 username=='test' 且 password=='123456' 时返回 200（唯一合法凭据，见 requestBody schema 的 example，正常用例必须使用该值）；\n2) 请求体缺少 username/password、字段为空字符串、字段类型不是字符串、password 长度超过 128，均返回 422（参数校验失败）；\n3) 字段齐全且格式合法、但用户名或密码不匹配（含任意非 test/123456 的合法字符串），返回 401（鉴权失败），此类场景不得标注为 422。",
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "登录成功，返回 token",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "凭据格式合法，但用户名或密码错误"
          },
          "422": {
            "description": "参数校验失败：字段缺失/空串/类型错误/密码超长"
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "summary": "Get User",
        "operationId": "get_user_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "用户名，非空字符串"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "title": "Password",
            "description": "密码，非空，长度不超过 128"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "example": {
          "password": "123456",
          "username": "test"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}